refactor: share wallet transaction helpers
