+++ src/components/DRepDelegate/index.js
@@ -11,6 +11,13 @@ import {
   delegateVote,
 } from "@site/src/utils/cardano/wallet";
 import drepAvatarsManifest from "@site/src/data/drep-avatars.json";
+import {
+  EXPECTED_NETWORK_ID,
+  EXPLORER_TX_BASE,
+  shortAddress,
+  stringifyError,
+  classifyError,
+} from "@site/src/utils/walletTx";
 import styles from "./styles.module.css";
 
 const AVATAR_SET = new Set(drepAvatarsManifest.ids);
@@ -20,8 +27,6 @@ const VP_MAX_LOVELACE = 50_000_000_000_000;  // 50M ada
 const DISPLAY_COUNT = 8;
 // data.cardano.org proxy caps POST bodies at 5120 bytes — ~80 drep_ids max per batch.
 const BATCH_SIZE = 50;
-const EXPECTED_NETWORK_ID = 1; // mainnet
-const EXPLORER_TX_BASE = "https://explorer.cardano.org/transaction/";
 const POOL_CACHE_KEY = "cardano-org.drep-pool.v2";
 const POOL_CACHE_TTL_MS = 15 * 60 * 1000; // 15 minutes
 
@@ -115,36 +120,6 @@ function isValidDRepId(input) {
   if (/^drep(_script)?1[a-z0-9]{40,}$/.test(t)) return true;
   if (/^[0-9a-f]{56,64}$/i.test(t)) return true;
   return false;
-}
-
-function shortAddress(addr) {
-  if (!addr) return "";
-  return `${addr.slice(0, 12)}…${addr.slice(-8)}`;
-}
-
-function stringifyError(err) {
-  if (!err) return "";
-  if (typeof err === "string") return err;
-  if (err instanceof Error) return err.message || err.toString();
-  const parts = [];
-  if (err.message) parts.push(String(err.message));
-  if (err.info) parts.push(String(err.info));
-  if (err.code != null) parts.push(`code=${err.code}`);
-  if (!parts.length) {
-    try { return JSON.stringify(err); } catch { return String(err); }
-  }
-  return parts.join(" · ");
-}
-
-function classifyError(err) {
-  const msg = stringifyError(err);
-  if (/StakeKeyNotRegistered|StakeNotRegistered/i.test(msg)) return "stakeNotRegistered";
-  // CIP-30 signTx throws code 2 for UserDeclined; submitTx code 2 means Failure.
-  // Only classify as cancel when the message corroborates it.
-  if (/user\s*(declined|rejected|cancel)|declined\s*by\s*user|rejected\s*by\s*user/i.test(msg)) {
-    return "userCancelled";
-  }
-  return "generic";
 }
 
 function WalletPicker({ onConnect, busy }) {
